+++ extender-api/src/main/kotlin/info/pzss/zomboid/extender/api/event/world/OnCreateRandomizedDeadSurvivors.kt
@@ -1,0 +1,6 @@
+package info.pzss.zomboid.extender.api.event.world
+
+import info.pzss.zomboid.extender.api.event.ZomboidScriptEvent
+import zombie.randomizedWorld.randomizedDeadSurvivor.RandomizedDeadSurvivorBase
+
+class OnCreateRandomizedDeadSurvivors(val entries: MutableList<RandomizedDeadSurvivorBase>) : ZomboidScriptEvent